pyproject.toml
[project]
name = "ise-api-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask==2.3.2",
    "logging==0.4.9.6",
    "python-dotenv==1.0.0",
    "requests==2.31.0",
    "urllib3==1.26.15",
]
